Do some work on project creation.

+++ app/src/main/java/com/example/cs446_group8/ui/add_project/AddProjectContract.java
@@ -6,7 +6,7 @@ import com.example.cs446_group8.ui.BaseContract;
 public interface AddProjectContract extends BaseContract {
 
     interface Presenter extends BaseContract.Presenter {
-        void saveButtonClicked(int projectId);
+        void saveButtonClicked(long projectId);
     }
 
 }
