+++ app/src/main/java/com/example/cs446_group8/ui/project_details/ProjectDetailsContract.java
@@ -5,9 +5,9 @@ import com.example.cs446_group8.ui.BaseContract;
 public interface ProjectDetailsContract extends BaseContract {
 
     interface Presenter extends BaseContract.Presenter {
-        void addButtonClicked(int projectId);
-        void settingsButtonClicked(int projectId);
-        void headcountsButtonClicked(int projectId);
+        void addButtonClicked(long projectId);
+        void settingsButtonClicked(long projectId);
+        void headcountsButtonClicked(long projectId);
         void scheduleButtonClicked();
     }
