+++ app/src/main/java/com/example/cs446_group8/ui/project_details/add_crop/head_count/MonthlyHeadCountContract.java
@@ -29,7 +29,7 @@ public interface MonthlyHeadCountContract extends BaseContract {
          */
         void changedHeadCount(int month, int headCount);
 
-        void saveButtonClicked(int projectId, String fromActivity);
+        void saveButtonClicked(long projectId, String fromActivity);
     }
 
 }
